save before removing controllers

+++ Todo/Dto/BlogPatchDto.cs
@@ -5,14 +5,14 @@ namespace Todo.Dto;
 public class BlogPatchDto
 {
      [Key]
-     public long Id { get; set; }
+     public int Id { get; set; }
      
      [Required]
      [StringLength(50, MinimumLength = 1)]
      public string Title { get; set; }
      
      [Required]
-     public long AuthorId { get; set; }
+     public int AuthorId { get; set; }
      
      [Required]
      [StringLength(250, MinimumLength = 1)]
